+++ example/lib/example.dart
@@ -3,15 +3,21 @@
 import 'package:nebula_dart_gdbc/nebula_dart_gdbc.dart';
 
 void main(List<String> args) async {
+  // 注册驱动
   DriverManager.registerDriver(NgDriver());
-
+  // 获取连接
   var conn = await DriverManager.getConnection(
-    'gdbc.nebula://127.0.0.1:9669/?username=root&password=nebula&space=test&timeout=1000',
+    'gdbc.nebula://127.0.0.1:9669/?space=test',
+    username: 'root', // username is optional
+    password: 'nebula', // password is optional
   );
 
+  // 创建语句
   var stmt = await conn.createStatement();
-  var rs = await stmt.executeQuery('SHOW SPACES;');
-
+  // 执行语句
+  var rs = await stmt.executeQuery(gql: 'SHOW SPACES;');
+  // 打印结果
   print(rs);
-  await conn.close();
+  // 关闭连接
+  conn.close();
 }