+++ lib/src/ng_result_handler.dart
@@ -4,12 +4,6 @@
 
 part of nebula_dart_gdbc;
 
-extension Int8ListToString on Int8List {
-  String utf8String() {
-    return utf8.decode(Uint8List.fromList(this));
-  }
-}
-
 NgResultSet handleResult(ng.ExecutionResponse rs, int? timezoneOffset) {
   ValueMetaData? meta = ValueMetaData();
   var rows = _handleValue(rs.data, meta, timezoneOffset);