+++ test/nebula_dart_gdbc_test.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2023- All nebula_dart_gdbc authors. All rights reserved.
+//
+// This source code is licensed under Apache 2.0 License.
+
 import 'dart:math';
 
 import 'package:nebula_dart_gdbc/nebula_dart_gdbc.dart';
@@ -25,20 +29,19 @@ void main() async {
   var spaceNamePrefix = 'my_space_';
 
   test('test create space', () async {
-    await stmt.executeQuery('use test;');
     await stmt.executeQuery(
-      '''
+      gql: '''
       CREATE SPACE IF NOT EXISTS $spaceNamePrefix$spaceKey (vid_type=FIXED_STRING(30));
       ''',
     );
 
     await Future.delayed(Duration(seconds: 10), () async {
-      await stmt.executeQuery('use $spaceNamePrefix$spaceKey;');
+      await stmt.executeQuery(gql: 'use $spaceNamePrefix$spaceKey;');
     });
   });
 
   test('Test List', () async {
-    var rs = await stmt.executeQuery('RETURN [1, 2, 3] as arr;');
+    var rs = await stmt.executeQuery(gql: 'RETURN [1, 2, 3] as arr;');
     expect(rs.value([0]), [1, 2, 3]);
     var firstMeta = rs.meta([0]);
     expect(firstMeta?.type, GdbTypes.list);
@@ -47,7 +50,7 @@ void main() async {
 
   tearDownAll(() async {
     await stmt.executeQuery(
-      '''
+      gql: '''
       DROP SPACE $spaceNamePrefix$spaceKey;
       ''',
     );